+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- IssuingBillingInvoice resource
 ### Changed
 - url attribute to IssuingToken resource
 - installmentCount attribute to IssuingPurchase resource